Fix: keep go-template schemas out of client bundle.
Split Zod schemas into go_template_types.ts so the hook does not import
server-side file processing code (and Node-only deps like tmp) in Vite.
